Check whether movement data frame contains valid tracker information

Checks the momentary speed, distance from reference position and/or distance from reference movement data frame. To compute speed, and distance across two data frames, the positions in the movement data frame(s) are first resampled to 10 Hz and then median-filtered over 3 data points, to avoid an impact of very short tracker glitches. Distance with respect to fixed position is checked for the first 200 ms of the tracking and operates on raw position values.

check_valid_movement(
  df,
  ref_df = NULL,
  ref_pos = NULL,
  max_dist = 2,
  max_speed = 10
)

Arguments

df

A movement data frame

ref_df

A reference movement data frame (default: NULL)

ref_pos

A reference initial position (list containing `pos_x` and `pos_z` items) (default: NULL)

max_dist

A maximum distance in metre (default: 2)

max_speed

A maximum speed in m/s (default: 10)

Value

A list of validity checks. If df is empty then the summary will be valid