Resample movement data after threat appears for all trials

For a particular datatype (e.g. 'head', 'threat', ...), resamples respective movement data to the desired sampling rate, from the moment the threat appears up the maximum trial duration from all trials. Non-existing time points for any trial are filled with "NA". Returns new movement data frames that have the same size for all trials.

resample_df_from_threat(df, datatype, sr)

Arguments

df
A trial results data frame. Expected columns: 'threat_appear_time', 'end_time', respective movement column

datatype
A string with the movement data type (e.g. 'head', 'threat', ...)

sr
Desired sample rate

Value

The data frame with a new column of resampled movements added, column name prepended with "resampled_"