Summarize fruit picking between a `min_time` and `max_time`. You can supply timestamps to `min_time` and `max_time` (e.g timestamp of begin escape) to average over only a period. In this case, you can also retrieve the collection rate by setting `rate` to 1.

NOTE: the fruit collection data frames contain no information on when the trial ends; only on the last event of the task. Hence it is not possible to extract the fruit collection rate without providing start and end time stamps.

summarise_fruit_task(df, min_time = 0, max_time = Inf, rate = FALSE)

Arguments

df
A fruit collection data frame. Expected columns are `time` and `event`.

min_time
minimum time in s (default trial start)

max_time
maximum time in s (default trial end)

rate
whether to extract rate (TRUE) or sum (FALSE, default)

Value

Sum of fruits or average of fruits picked per second, between `min_time` and `max_time`.